rag-mongodb-atlas/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# MongoDB
pymongo==4.6.1
motor==3.3.2

# Environment Management
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# LLM Integration
groq==0.4.2

# Embeddings & ML
sentence-transformers==2.3.1
transformers==4.37.2
torch==2.1.2
pillow==10.2.0
numpy==1.26.3

# Vector Search
faiss-cpu==1.7.4

# Utilities
requests==2.31.0
httpx==0.26.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Logging
loguru==0.7.2
